+++ express-server/src/database/config.ts
@@ -1,4 +1,4 @@
-const mongoose = require('mongoose');
+import * as mongoose from 'mongoose';
 import { CONFIG } from '../config';
 
 const mongoUri = CONFIG.mongoUri;
@@ -11,7 +11,7 @@ mongoose.connect(mongoUri, {
     .then(() => console.log("Database Connected"))
     .catch((err) => console.log(err));
 
-mongoose.Promise = global.Promise;
+(mongoose as any).Promise = global.Promise;
 
 require('./logInfo.model');
 require('./user.model');