+++ express-server/src/lib/definitions.ts
@@ -1,0 +1,5 @@
+import * as express from 'express';
+
+export interface IGetUserAuthInfoRequest extends express.Request {
+    user: any // or any other type
+}