+++ express-server/src/lib/utils.ts
@@ -1,7 +1,7 @@
 // reference: https://medium.com/swlh/everything-you-need-to-know-about-the-passport-jwt-passport-js-strategy-8b69f39014b0
-const jsonwebtoken = require('jsonwebtoken');
 import * as fs from "fs";
 import * as path from "path";
+import * as jsonwebtoken from 'jsonwebtoken';
 
 const pathToKey = path.join(__dirname, '../..', 'id_rsa_priv.pem');
 const PRIV_KEY = fs.readFileSync(pathToKey, 'utf8');
@@ -9,7 +9,7 @@ const PRIV_KEY = fs.readFileSync(pathToKey, 'utf8');
 /**
  * @param {*} user - The user object.  We need this to set the JWT `sub` payload property to the MongoDB user ID
  */
-function issueJWT(user) {
+ export function issueJWT(user) {
   const _id = user._id;
 
   const expiresIn = '1d';
@@ -26,5 +26,3 @@ function issueJWT(user) {
     expires: expiresIn
   }
 }
-
-module.exports.issueJWT = issueJWT;