+++ express-server/src/routes/auth.ts
@@ -1,25 +1,28 @@
-const router = require('express').Router();
-const passport = require('passport');
-const querystring = require("querystring");
-const User = require('mongoose').model('User');
+import * as express from 'express';
+import * as passport from 'passport';
+import * as mongoose from 'mongoose';
+const router = express.Router();
+const User = mongoose.model('User');
 
 import { CONFIG } from '../config';
-const utils = require('../lib/utils');
+import * as utils from '../lib/utils';
+import { IGetUserAuthInfoRequest } from '../lib/definitions';
 
-router.get('/protected', passport.authenticate('jwt', { session: false }), async (req, res, next) => {
-  console.log("--protected");
+router.get('/protected', passport.authenticate('jwt', { session: false }),
+  async (req: IGetUserAuthInfoRequest, res, next) => {
+    console.log("--protected");
 
-  const user = await User.findOne({ _id: req.user._id })
-    .exec()
-    .then((user) => {
-      return user;
-    })
-    .catch((err) => {
-      return 'error occured';
-    });
+    const user = await User.findOne({ _id: req.user._id })
+      .exec()
+      .then((user) => {
+        return user;
+      })
+      .catch((err) => {
+        return 'error occured';
+      });
 
-  res.status(200).json(user);
-});
+    res.status(200).json(user);
+  });
 router.get('/test', (req, res, next) => {
   res.status(200).json({ success: true, msg: "You are successfully authenticated to this route!" });
 });
@@ -32,7 +35,7 @@ router.get('/google',
 
 router.get('/' + CONFIG.oAuth_redirect_postfix,
   passport.authenticate('google', { session: false }),
-  function (req, res) {
+  function (req: IGetUserAuthInfoRequest, res) {
 
     const tokenObject = utils.issueJWT(req.user);
     const data = { success: true, token: tokenObject.token, expiresIn: tokenObject.expires };